Remove unused import of Apache Commons Lang
Commit efb2bf4b3469d4eab136be2c94df36c56ad1d70d reverted the removal of
the compile-only Apache Commons Lang in commit
4a4bd30be8b792670912d812c9b466b1aa7781bf, as it was no longer relevant
to this branch. However, that commit had also removed the runtime import
of the latest version of Commons Lang from
9d90db2be830287ad0c06545cb3d763c73a462cd, which should have been
retained, since all the usages of the library in that commit had been
removed at that point. This commit reinstates the removal of this
import.