Implement a specific parser for the version scheme
There is a custom version scheme used in the app, which can be defined
as 'X.X.X.Y', where X represents numbers, and Y represents an optional
alphanumeric string. Y is ignored for comparison purposes, since it
should always coincide with a change in the X portion.

A custom Version class has been written to stricly represent the app
scheme, replacing the lenient ComparableVersion and
DefaultArtifactVersion that were imported from the Maven Artifact
project. A test suite for it has been written as well. It simply splits
the string by the period token, takes the first three (or less)
substrings it finds, and converts them into numbers. If any of them
can't be converted into an unsigned integer, then it throws a
NumberFormatException.

The NewVersionBroadcastReceiver terminates it's operations if it
encounters a NumberFormatException while parsing the version from the
new version header data.